feat-delete additional folder
