--- ExampleService/ExampleService/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "dotnet.defaultSolution": "ExampleService.sln"
-}
