+++ ExampleService/ExampleService_WebApi/Controllers/TaskController.cs
@@ -5,21 +5,37 @@ namespace ExampleService_WebApi;
 [ApiController]
 [Route("[controller]")]
 public class TaskController :  ControllerBase
-{   
+{
+    #region PROPERTIES
     private readonly ITaskInterface _interface;
+
+    #endregion
+
+    #region CONSTRUCTOR
     public TaskController(ITaskInterface interfaceClass)
     {
         _interface= interfaceClass;
     }
-    
+
+    #endregion
+
+    #region CRUD LOGICS
+
     /// <summary>
     /// to get the list of task
     /// </summary>
     /// <returns>List of task avail in memory</returns>
+    /// 
+
     [HttpGet("List of your Task")]
     public IActionResult Get()
     {   
-        var taskList = _interface.ListofTask();
+        List<TaskModel> taskList = _interface.ListofTask();
+
+        if(taskList == null || taskList.ToList().Count == 0 ) 
+        { 
+            return NotFound("No Task Found");
+        }
         
         List<TaskDTO> outputTaskDTO = new List<TaskDTO>();    
         foreach(var tasks in taskList)
@@ -39,9 +55,15 @@ public class TaskController :  ControllerBase
    /// </summary>
    /// <param name="task">Ojects to be added</param>
    /// <returns>Success message of adding task</returns>
+   /// 
+
     [HttpPost("Create your Task here")]
     public IActionResult Post([FromBody] TaskDTO taskRequest)
-    {         
+    {     
+        if(taskRequest == null)
+        {
+            return BadRequest("Provide Data to be added");
+        }
         var taskModel = new TaskModel();
 
         taskModel.TaskStatus = taskRequest.TaskStatus;
@@ -53,4 +75,5 @@ public class TaskController :  ControllerBase
         return Ok("Task added successfully");
     }
 
+    #endregion
 }
